+++ src/TableParser/HTML/HTMLStringTableParser.php
@@ -7,9 +7,13 @@
 namespace Ingenerator\BehatTableAssert\TableParser\HTML;
 
 
+use Behat\Gherkin\Node\TableNode;
+use Dom\HTMLDocument;
+use Dom\HTMLElement;
 use Ingenerator\BehatTableAssert\TableNode\PaddedTableNode;
 use LibXMLError;
 use Masterminds\HTML5;
+use function Dom\import_simplexml;
 
 /**
  * Parses an HTML string for a <table> element into a TableNode. The table must have a single row
@@ -90,33 +94,43 @@ class HTMLStringTableParser
      */
     protected function parseHTMLString($html)
     {
-        $old_use_internal_errors = \libxml_use_internal_errors(TRUE);
         try {
-            $html5 = new HTML5();
-            $dom = $html5->loadHTML(
-              '<!DOCTYPE html><html>'
-              .'<head><meta http-equiv="Content-Type" content="text/html; charset=utf-8"></head>'
-              .'<body>'.\trim($html).'</body>'
-              .'</html>'
+            set_error_handler(
+                fn($errno, $errstr) => throw new \InvalidArgumentException(
+                    sprintf("Invalid HTML: %s\n\n===HTML===\n%s", $errstr, $html),
+                ),
             );
 
-            $table_elem = $dom->getElementsByTagName('body')->item(0)->firstChild;
-            $table      = \simplexml_import_dom($table_elem);
-            if ($errors = \libxml_get_errors()) {
-                $this->throwInvalidHTMLException($html, $errors);
-            }
+            $dom = HTMLDocument::createFromString(
+                sprintf(
+                    <<<'HTML'
+                    <!DOCTYPE html>
+                    <html>
+                    <head><meta http-equiv="Content-Type" content="text/html; charset=utf-8"></head>
+                    <body>%s</body>
+                    </html>
+                    HTML,
+                    trim($html),
+                ),
+            );
         } finally {
-            \libxml_clear_errors();
-            \libxml_use_internal_errors($old_use_internal_errors);
-
+            restore_error_handler();
         }
+
+        $table_elem = $dom->getElementsByTagName('body')->item(0)->firstChild;
+
+        if (!$table_elem instanceof HTMLElement) {
+            throw new \InvalidArgumentException(
+                sprintf("Expected html root element but got %s\n\n===HTML===\n%s", get_debug_type($table_elem), $html),
+            );
+        }
+        $table = \simplexml_import_dom($table_elem);
 
         return $table;
     }
 
     /**
-     * @param string        $html
-     * @param LibXMLError[] $errors
+     * @deprecated no longer called by the library
      */
     protected function throwInvalidHTMLException($html, $errors)
     {
@@ -224,7 +238,7 @@ class HTMLStringTableParser
      */
     protected function parseCellText(\SimpleXmlElement $cell)
     {
-        $text = \trim(\preg_replace('/\s+/', ' ', \dom_import_simplexml($cell)->textContent));
+        $text = \trim(\preg_replace('/\s+/', ' ', import_simplexml($cell)->textContent));
 
         if ($prefix = (string) $cell['data-behat-table-prefix']) {
             $text = $prefix.' '.$text;