+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* Use native `HTMLDocument` parsing instead of masterminds/html5
+* Drop support for PHP < 8.4 (to allow us to use native HTML parsing)
+
+## 1.3.3 / 2026-03-10
+
+* Fix handling and reporting parse errors when parsing an html table.
+
 ## 1.3.2 / 2025-05-06
 
 * Support PHP 8.4 (thanks @mharmuth)